DB schema: decimal precision audit across remaining entities (Phase 3 F8-broad)
WU-10 (F8-partial) covered PO line + SO line + Customer.creditLimit. This
sweep covers the rest of the int->decimal pattern across the data model:

  Quantities (decimal(18,4)): InvoiceLine, QuoteLine, ShipmentLine,
    RecurringOrderLine, LotRecord, ReceivingRecord, AutoPoSuggestion.
  Aggregates (decimal): ShippingSummaryReportItem.TotalItems,
    SalesOrderShipmentLineModel.Quantity.

Money fields already at appropriate precision were left alone (Payment.Amount
already decimal(18,4); Expense / RecurringExpense.Amount already decimal(18,2)).
Auto-PO whole-unit rounding behavior preserved per scope-creep rule.

Single migration: WU23_DecimalPrecisionBroad_Audit. Existing int values cast
cleanly to decimal(18,4); zero data loss. End-to-end fractional-quantity
round-trip verified: invoice (2.75 qty x 4.50 = 12.375), quote (3.125 qty),
lot (12.5 qty), receive (2.25 against 10.5 PO line) all preserve fractional
values across the full quote-to-cash and receive-to-stock cycles. SO line
fractional qty no longer truncates when invoiced/shipped/lot-traced.

Cases addressed: F8-broad (TODO from phase-3-todos.md). Closes WU-10's
flagged silent-corruption follow-up on InvoiceLine/QuoteLine/ShipmentLine.

+++ qb-engineer.api/Features/DomainEvents/Handlers/OnSalesOrderConfirmed_CheckBomMaterials.cs
@@ -129,7 +129,9 @@ public class OnSalesOrderConfirmed_CheckBomMaterials(
                 {
                     PartId = materialPartId,
                     VendorId = vendorId,
-                    SuggestedQty = (int)Math.Ceiling(shortfall),
+                    // Phase 3 / WU-23 (F8-broad): AutoPoSuggestion.SuggestedQty is
+                    // now decimal; preserve fractional precision for UoM-aware demand.
+                    SuggestedQty = Math.Ceiling(shortfall),
                     NeededByDate = deliveryDate.AddDays(-7), // Need materials before delivery
                     SourceSalesOrderIds = so.Id.ToString(),
                     Status = AutoPoSuggestionStatus.Pending,
