+++ qb-engineer.core/Entities/QuoteLine.cs
@@ -5,7 +5,8 @@ public class QuoteLine : BaseEntity
     public int QuoteId { get; set; }
     public int? PartId { get; set; }
     public string Description { get; set; } = string.Empty;
-    public int Quantity { get; set; }
+    // Phase 3 / WU-23 (F8-broad): UoM-aware fractional quantities.
+    public decimal Quantity { get; set; }
     public decimal UnitPrice { get; set; }
     public int LineNumber { get; set; }
     public string? Notes { get; set; }
