+++ qb-engineer.core/Entities/ReceivingRecord.cs
@@ -5,7 +5,9 @@ namespace QBEngineer.Core.Entities;
 public class ReceivingRecord : BaseAuditableEntity
 {
     public int PurchaseOrderLineId { get; set; }
-    public int QuantityReceived { get; set; }
+    // Phase 3 / WU-23 (F8-broad): receiving against a fractional-quantity PO
+    // line preserves precision (PO line itself was promoted in WU-10).
+    public decimal QuantityReceived { get; set; }
     public string? ReceivedBy { get; set; }
     public int? StorageLocationId { get; set; }
     public string? Notes { get; set; }
