+++ qb-engineer.core/Entities/RecurringOrderLine.cs
@@ -5,7 +5,8 @@ public class RecurringOrderLine : BaseEntity
     public int RecurringOrderId { get; set; }
     public int PartId { get; set; }
     public string Description { get; set; } = string.Empty;
-    public int Quantity { get; set; }
+    // Phase 3 / WU-23 (F8-broad): UoM-aware fractional quantities.
+    public decimal Quantity { get; set; }
     public decimal UnitPrice { get; set; }
     public int LineNumber { get; set; }
