+++ qb-engineer.core/Models/InvoiceLineResponseModel.cs
@@ -5,7 +5,7 @@ public record InvoiceLineResponseModel(
     int? PartId,
     string? PartNumber,
     string Description,
-    int Quantity,
+    decimal Quantity,
     decimal UnitPrice,
     decimal LineTotal,
     int LineNumber);
