+++ qb-engineer.core/Models/ReceiveItemsRequestModel.cs
@@ -1,4 +1,4 @@
 namespace QBEngineer.Core.Models;
 
 public record ReceiveItemsRequestModel(List<ReceiveLineModel> Lines);
-public record ReceiveLineModel(int LineId, int Quantity, int? StorageLocationId, string? Notes);
+public record ReceiveLineModel(int LineId, decimal Quantity, int? StorageLocationId, string? Notes);
