+++ qb-engineer.core/Models/ShippingSummaryReportItem.cs
@@ -3,6 +3,7 @@ namespace QBEngineer.Core.Models;
 public record ShippingSummaryReportItem(
     string Status,
     int Count,
-    int TotalItems,
+    // Phase 3 / WU-23 (F8-broad): aggregate of ShipmentLine.Quantity (now decimal).
+    decimal TotalItems,
     int OnTimeCount,
     int LateCount);
